requirements.txt
alembic==1.8.1
atomicwrites==1.4.1
attrs==21.4.0
Babel==2.10.3
click==8.1.3
colorama==0.4.4
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
importlib-resources==5.9.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
py==1.11.0
PyJWT==1.7.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
ruamel.yaml==0.16.5
six==1.16.0
SQLAlchemy==1.4.36
tomli==2.0.1
Werkzeug==2.1.2
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
zipp==3.8.0
